added sqlite database, saving games to it and pausing games.

+++ BlackJackLogic/Course.cs
@@ -5,12 +5,12 @@ namespace BlackJackLogic
     {
         public bool IsWin { get; set; }
         public float _CourseValue {  get; set; }
-        public float PlayerBet {  get; set; }
+        public int PlayerBet {  get; set; }
         public Course()
         {
             IsWin = false;
             _CourseValue = 1.33f;
-            PlayerBet = 0f;
+            PlayerBet = 0;
         }
     }
 }
